Move index file
